requirements.txt
appdirs==1.4.4
asgiref==3.3.2
astroid==2.4.2
async-timeout==4.0.2
boto3==1.26.3
botocore==1.29.3
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.4
ConfigArgParse==1.5.3
cryptography==3.3.2
cycler==0.10.0
Deprecated==1.2.13
distlib==0.3.1
Django==3.2.10
django-admin-tools==0.9.1
django-audittrail==0.2.1
django-auto-prefetch==0.0.6
django-cors-headers==3.7.0
django-filter==2.4.0
django-passwordless-auth==2020.12.3
django-redis==5.2.0
django-reversion==3.0.9
django-soft-delete==0.9.18
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-api-key==2.1.0
djangorestframework-jsonapi==5.0.0
docker==6.1.3
drf-expiring-token==1.1.2
drf-flex-fields==0.8.8
dynamic-rest==2.1.1
et-xmlfile==1.0.1
Faker==5.0.1
filelock==3.0.12
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.12.0
geventhttpclient==2.0
ghost==0.6.1
greenlet==1.1.2
gunicorn==20.1.0
hashids==1.3.1
idna==2.10
importlib-metadata==4.12.0
inflection==0.5.1
isort==5.7.0
itsdangerous==2.1.2
jdcal==1.4.1
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
locust==2.10.2
MarkupSafe==2.1.1
matplotlib==3.3.4
mccabe==0.6.1
msgpack==1.0.4
numpy==1.22.0
openpyxl==3.0.6
packaging==21.3
pandas==1.2.0
Pillow==9.3.0
psutil==5.9.1
psycopg2==2.8.6
pycparser==2.21
pycryptodome==3.10.1
Pygments==2.7.4
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.1
python-dotenv==0.19.0
python-json-logger==2.0.4
pytz==2020.5
pywin32==306
pyzmq==22.3.0
redis==4.3.4
requests==2.31.0
roundrobin==0.0.4
s3transfer==0.6.0
selenium==3.141.0
sentry-sdk==1.4.3
serpy==0.3.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
tinydb==4.3.0
toml==0.10.2
typing_extensions==4.3.0
tzdata==2022.6
urllib3==1.26.4
virtualenv==20.0.33
websocket-client==1.6.4
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.12.1
xlrd==2.0.1
XlsxWriter==3.0.3
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
